+++ python/turbodbc_test/test_options.py
@@ -15,7 +15,8 @@ def test_options_with_overrides():
                            autocommit=True,
                            large_decimals_as_64_bit_types=True,
                            limit_varchar_results_to_max=True,
-                           force_extra_capacity_for_unicode=True)
+                           force_extra_capacity_for_unicode=True,
+                           fetch_wchar_as_char=True)
 
     assert options.read_buffer_size.rows == 123
     assert options.parameter_sets_to_buffer == 2500
@@ -26,3 +27,4 @@ def test_options_with_overrides():
     assert options.large_decimals_as_64_bit_types == True
     assert options.limit_varchar_results_to_max == True
     assert options.force_extra_capacity_for_unicode == True
+    assert options.fetch_wchar_as_char = True